test: mapped more pool configurator-related tests
